requirements.txt
python-dotenv~=1.1.1
telebot~=0.0.5
pyTelegramBotAPI~=4.27.0
tinydb~=4.8.2
APScheduler~=3.10.4
pytz~=2025.2
jdatetime~=5.2.0
reportlab~=4.4.3
pymongo~=4.13.2